PMXモデルのマテリアル情報を一部追加して描画した

+++ IQByDirectX12/IQByDirectX12/Model/PMDModelData.cpp
@@ -7,7 +7,7 @@
 #include "InstancingDataManager.h"
 
 PMDModelData::PMDModelData(ComPtr<ID3D12Device> device, const std::vector<PMDVertex>& vertexData, const std::vector<unsigned short>& indexData, const std::vector<PMDMaterial>& materials)
-	: ModelData(VertexBuffer::Create(device, (void*)vertexData.data(), vertexData.size(), sizeof(PMDVertex)), IndexBuffer::Create(device, (void*)indexData.data(), indexData.size(), sizeof(short)), DescriptorHeap::Create(device, 1 + materials.size()) )
+	: ModelData(VertexBuffer::Create(device, (void*)vertexData.data(), vertexData.size(), sizeof(PMDVertex)), IndexBuffer::Create(device, (void*)indexData.data(), indexData.size(), sizeof(short)), DescriptorHeap::Create(device, 1 + (int)materials.size()) )
 {
 	SetVertexData(vertexData);
 	SetIndexData(indexData);
